+++ app/Http/Controllers/ProjectController.php
@@ -43,7 +43,9 @@ class ProjectController extends Controller
       if ($request->file('image')) {
         $image = $request->file('image');
         $imagename = time().$image->hashname();
-        Storage::delete(['public/images/projects/thumbnails/'.$project->image,'public/images/projects/originals/'.$project->image,'public/images/projects/mediums/'.$project->image,]);
+        if ($project->image!='card-1.jpg' && $project->image!='card-2.jpg' && $project->image!='card-3.jpg'){
+          Storage::delete(['public/images/projects/thumbnails/'.$project->image,'public/images/projects/originals/'.$project->image,'public/images/projects/mediums/'.$project->image,]);
+        }
         $image->storeAs('public/images/projects/originals/', $imagename);
         $resized = ImgInt::make($image)->resize(350,260)->save();
         Storage::put('public/images/projects/mediums/'.$imagename, $resized);
